api/pyproject.toml
[tool.poetry]
name = "projeto2"
version = "0.1.0"
description = ""
authors = ["Vlademir Santos <vlademir.santos@nubank.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.1"
uvicorn = "^0.23.1"
requests = "^2.31.0"
aiohttp = "^3.8.5"
sqlalchemy = "^2.0.19"
pytest = "^7.4.0"
alembic = "^1.11.1"
python-decouple = "^3.8"
httpx = "^0.24.1"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
fastapi-pagination = "^0.12.6"
mysql-connector-python = "^8.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
